Add configurable response header including Request ID.
The Request-Tracker ID that's made available in the logging context is
very useful for debugging requests for a troubled client.  By adding the
same request id as a response header, we can easily ask the same client
for a request id that failed to narrow down our search.

To ensure functionality doesn't change for existing clients, the request
header must be explicitly enabled before it will be added.  The same
header name that's used in the request object will be used.